backend/requirements.txt
Flask
Flask-Cors
python-dotenv
cloudinary
pymongo
